pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "coilshield-iccp"
version = "0.1.0"
description = "CoilShield ICCP — Raspberry Pi HVAC coil cathodic protection controller"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "RPi.GPIO>=0.7.0",
    "flask>=3.0.0",
    "textual>=1.0.0",
    "pytest>=7.0",
    "smbus2>=0.4.0",
    "pi-ina219>=1.4.0",
    "adafruit-circuitpython-ina3221>=1.0.0",
]

[project.optional-dependencies]
ble = [
    "dbus-python>=1.3.2",
    "PyGObject>=3.42.0",
]
cloud = [
    "paho-mqtt>=1.6.3",
    "inotify-simple>=2.0; sys_platform == 'linux'",
]
supabase = [
    "supabase>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
iccp = "iccp_cli:main"
iccp-ble-provision = "pi_edge.ble_provision:main"
iccp-cloud-register = "pi_edge.cloud_register:main"
iccp-commission-mqtt = "pi_edge.commission_mqtt_bridge:main"
iccp-telemetry-mqtt = "pi_edge.telemetry_mqtt_sidecar:main"
iccp-edge-doctor = "pi_edge.edge_doctor:main"

[tool.setuptools.package-dir]
"" = "src"
"config" = "config"

[tool.setuptools.packages.find]
where = ["src", "."]
include = ["pi_edge*", "config*", "generated*"]

[tool.setuptools]
py-modules = [
    "main",
    "cli_events",
    "console_ui",
    "telemetry_queries",
    "iccp_runtime",
    "diagnostics",
    "iccp_cli",
    "hw_probe",
    "i2c_bench",
    "control",
    "sensors",
    "reference",
    "commissioning",
    "ble_provisioning",
    "platform_util",
    "logger",
    "dashboard",
    "leds",
    "temp",
    "tui",
    "polarization_safety",
    "channel_labels",
    "iccp_electrolyte",
    "cloud_sync",
    "tech_api",
]
